more resilient downloading
Former-commit-id: 70aa3c93827a9f026c16dda454bd40728c1b96f5
Former-commit-id: 188d030361004d6608924f52e7018e7e99692225

+++ packages/fetch-engine/src/download.ts
@@ -61,13 +61,16 @@ export async function download(options: DownloadOptions): Promise<BinaryPaths> {
       options.binaries['introspection-engine']
     : null
 
-  if (baseDir) {
+  const downloadDoneFile = baseDir ? path.join(baseDir, 'download-done') : null
+  const nativeDownloadsDone = downloadDoneFile ? fs.existsSync(downloadDoneFile) : false
+
+  if (baseDir && !nativeDownloadsDone) {
     try {
       const writeTestPath = path.join(baseDir, 'write-test')
       fs.writeFileSync(writeTestPath, 'write-test')
       fs.unlinkSync(writeTestPath)
     } catch (e) {
-      if (options.failSilent) {
+      if (options.failSilent || e.code !== 'EACCES') {
         return
       } else {
         throw new Error(`Can't write to ${baseDir} please make sure you install "prisma2" with the right permissions.`)
@@ -75,8 +78,6 @@ export async function download(options: DownloadOptions): Promise<BinaryPaths> {
     }
   }
 
-  const downloadDoneFile = baseDir ? path.join(baseDir, 'download-done') : null
-  const nativeDownloadsDone = downloadDoneFile ? fs.existsSync(downloadDoneFile) : false
   const everythingDownloaded =
     nativeDownloadsDone &&
     (!options.binaryTargets || (options.binaryTargets.length === 1 && options.binaryTargets[0] === platform))
